Move test to the correct test module
